Log time so far, remove home module
Home module was useless so deleted.
Spent 1.5hrs (close to 2) learning the differences between TypeScript and JavaScript.
Although tsx seems much faster and more powerful, JS is simpler to learn as beginner web dev, and NextJS (framework
I'm using) is built on tsx so should still be fast.
If this app gets scaled to a huge app the conversion to TypeScript is best to happen sooner rather than later to
maintain effective speed and scalability management.