Correct syntax of CardContext file
Changed a wrong line trying to call a submethod of JSON.parse() as .localStorage,
however localStorage.getItem(<key>) is its own method to return an item with a given key.
From there I wanted to JSON.parse it
Also set up the key as 'mg-items' for my items and changed the != to !== syntax as this is JS standard.