Add curried function addToBasket and debugged issue
Issue with image not loading on its page (see previous commit)
to do with trying to dynamically allocate string to link when pushing with
router. Was using "'" instead of "`" and the ${variablename} inside meant
the line was simply '/${variablename}' instead of '/product_id<idnum>'

debugged using console logs of all of the constants set,
added a check for undefined product.